requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aleph-alpha-client==2.16.0
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
dataclasses-json==0.5.7
filelock==3.9.0
frozenlist==1.3.3
greenlet==2.0.2
huggingface-hub==0.12.1
idna==3.4
joblib==1.2.0
langchain==0.0.94
llama-index==0.4.14
lxml
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
openai==0.26.5
packaging==23.0
pandas==1.5.3
pycares==4.3.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.5
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests
six==1.16.0
SQLAlchemy==1.4.46
tenacity==8.1.0
tiktoken==0.2.0
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
python-dotenv==1.0.0
flask
html2text
beautifulsoup4
blinker
Scrapy==2.8.0
Twisted==22.10.0
attrs==22.2.0